Better implementation for previous patch in r2943.
git-svn-id: svn://svn.savannah.gnu.org/myserver/trunk@2944 c4c922ea-a70d-0410-937d-ec1daf91936c

+++ myserver/include/filter/stream.h
@@ -25,7 +25,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 typedef void* FileHandle;
 #endif
 #ifdef NOT_WIN
-typedef long  FileHandle;
+typedef int  FileHandle;
 #endif
 
 /*!
